refactor: Introduce new agent components and restructure imports
- Added new files for `parsers`, `providers`, `runtime`, `tokens`, and `tools` to enhance agent functionality.
- Updated import paths in `agent.py` to reflect the new structure.
- Implemented JSON parsing and model retrieval logic in the new components for improved agent operations.
